imgToAscii/backend/requirements.txt
flask
flask-cors
pillow
numpy
scikit-learn